packages/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.27
psycopg[binary]>=3.2
alembic==1.13.1
pydantic-settings==2.12.0
greenlet==3.3.0
python-multipart==0.0.20
loguru==0.7.3
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
email-validator==2.3.0
requests==2.31.0
daily-python==0.10.1
deepgram-sdk==3.2.0
google-generativeai==0.3.2
groq==0.11.0
pipecat-ai==0.0.90
livekit==0.16.0
livekit-agents==0.8.0
openai==1.54.0
google-cloud-texttospeech==2.16.0
websockets==12.0
